Fix output directory fallback and optimize conversion pipeline
